Move interactive gui code; #0188
