+++ wbfm/gui/interactive_folder_grid_plot_explorer.py
@@ -2,7 +2,7 @@ import argparse
 
 from wbfm.gui.utils.utils_gui import build_gui_for_grid_plots
 from wbfm.utils.projects.finished_project_data import ProjectData
-from wbfm.utils.visualization.plot_traces import ClickableGridPlot
+from wbfm.gui.utils.utils_gui_matplot import ClickableGridPlot
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='Build GUI for exploring folders of grid plots')
